replace js to tsx
